+++ implants/imix/src/main.rs
@@ -25,16 +25,8 @@ async fn run(config_path: String) -> Result<(), imix::Error> {
     let config_file = File::open(config_path)?;
     let config: imix::Config = serde_json::from_reader(config_file)?;
 
-
+    #[cfg(any(target_os = "windows", target_os = "linux", target_os = "macos"))]
     return Ok(imix::common::main_loop().await);
-
-    #[cfg(target_os = "windows")]
-        return imix::windows::run(config).await;
-
-    #[cfg(target_os = "linux")]
-        if Path::new(imix::linux::SYSTEMD_DIR).is_dir() {
-            return imix::linux::run(config).await;
-        }
 
     unimplemented!("The current OS/Manager is not supported")
 }
